Extract clipboard logic
