+++ src/editor.c
@@ -12,8 +12,7 @@
 #include "config.h"
 #include "files.h"
 #include "undo.h"
-
-#define CLIPBOARD_SIZE 4096
+#include "clipboard.h"
 
 int cursor_x = 1, cursor_y = 1;
 char current_filename[256] = "";  // Name of the current file being edited
@@ -22,12 +21,6 @@ int line_count = 0;  // Number of lines in the text buffer
 int start_line = 0;  // Index of the first visible line in the text window
 int runeditor = 0;  // Flag to control the main loop of the editor
 WINDOW *text_win;  // Pointer to the ncurses window for displaying the text
-
-// Global clipboard
-char *clipboard;
-bool selection_mode = false;
-int sel_start_x = 0, sel_start_y = 0;
-int sel_end_x = 0, sel_end_y = 0;
 
 // Undo and redo stacks
 Node *undo_stack = NULL;
@@ -283,89 +276,6 @@ static void initialize_key_mappings() {
     key_mappings[key_mapping_count++] = (KeyMapping){0, NULL}; /* terminator */
 }
 
-void start_selection_mode(int cursor_x, int cursor_y) {
-    selection_mode = true;
-    sel_start_x = cursor_x;
-    sel_start_y = cursor_y;
-    sel_end_x = cursor_x;
-    sel_end_y = cursor_y;
-}
-
-void end_selection_mode() {
-    selection_mode = false;
-    copy_selection();
-}
-
-void copy_selection() {
-    int start_x = sel_start_x;
-    int end_x = sel_end_x;
-    int start_y, end_y;
-    
-    if (sel_start_y < sel_end_y) {
-        start_y = sel_start_y;
-        end_y = sel_end_y;
-    } else {
-        start_y = sel_end_y;
-        end_y = sel_start_y;
-    }
-
-    if (clipboard == NULL) {
-        return;
-    }
-
-    clipboard[0] = '\0';  // Clear clipboard
-    for (int y = start_y; y <= end_y; y++) {
-        if (y == start_y) {
-            strncat(clipboard, &text_buffer[y - 1 + start_line][start_x - 1], end_x - start_x + 1);
-        } else {
-            strcat(clipboard, "\n");
-            strcat(clipboard, text_buffer[y - 1 + start_line]);
-        }
-    }
-}
-
-void paste_clipboard(int *cursor_x, int *cursor_y) {
-    if (clipboard == NULL) {
-        return;
-    }
-
-    /* Copy clipboard contents to a temporary buffer so that
-       tokenizing does not modify the original clipboard. */
-    char tmp[CLIPBOARD_SIZE];
-    strcpy(tmp, clipboard);
-
-    char *line = strtok(tmp, "\n");
-    while (line) {
-        int len = strlen(line);
-        memmove(&text_buffer[*cursor_y - 1 + start_line][*cursor_x - 1 + len], &text_buffer[*cursor_y - 1 + start_line][*cursor_x - 1], strlen(&text_buffer[*cursor_y - 1 + start_line][*cursor_x - 1]) + 1);
-        memcpy(&text_buffer[*cursor_y - 1 + start_line][*cursor_x - 1], line, len);
-        line = strtok(NULL, "\n");
-        (*cursor_y)++;
-        *cursor_x = 1;
-    }
-}
-
-void handle_selection_mode(int ch, int *cursor_x, int *cursor_y) {
-    if (ch == KEY_UP) {
-        if (*cursor_y > 1) (*cursor_y)--;
-        sel_end_y = *cursor_y;
-    }
-    if (ch == KEY_DOWN) {
-        if (*cursor_y < LINES - 4) (*cursor_y)++;
-        sel_end_y = *cursor_y;
-    }
-    if (ch == KEY_LEFT) {
-        if (*cursor_x > 1) (*cursor_x)--;
-        sel_end_x = *cursor_x;
-    }
-    if (ch == KEY_RIGHT) {
-        if (*cursor_x < COLS - 6) (*cursor_x)++;
-        sel_end_x = *cursor_x;
-    }
-    if (ch == 10) { // CTRL-J to end selection mode
-        end_selection_mode();
-    }
-}
 
 /**
  * Deletes the current line from the text buffer.
